All serial port tests are now passing in simulation
This took a lot of effort! The VICE cia-sdr-icr tests don't actually
test the contents of the ICR register after switching SP MODE from
OUTPUT to INPUT, and consequently emulators don't get this right.

A new test will be added to https://sourceforge.net/p/vice-emu/bugs/1219/

+++ gateware/cia_core_sim.cpp
@@ -91,7 +91,7 @@ static void parse_args(int argc, char** argv) {
                 if (tod_frequency < 1 || tod_frequency > 1000000) {
                     goto fail;
                 }
-                tod_timestep = 1e12/tod_frequency/2;
+                tod_timestep = uint64_t(1e12/tod_frequency/2);
             } catch (exception const&) {
                 goto fail;
             }
